dry the code
